% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_estimate.R
\docType{data}
\name{factor_estimate}
\alias{factor_estimate}
\title{The factor_estimate class provides an inheritance model for effective factor estimation approaches.
This class should be inherited from but not instanciated from (ie it is "abstract").}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
factor_estimate
}
\description{
TODO:
- Glue code with model_factor.
}
\keyword{datasets}

