% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_estimate_composite.R
\docType{class}
\name{factor_estimate_composite}
\alias{factor_estimate_composite}
\title{factor_estimate_composite}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
factor_estimate_composite
}
\value{
An instance of the factor_estimate_composite \code{\link{R6Class}}.
}
\description{
A risk model "composite factor estimate".
A "composite factor estimate" is a factor estimation method that is based on \code{n} sub-factors.
The sub-factors may in their turn be estimated with arbitrary estimation methods.
\code{factor_estimate_composite} is an "abstract" class that shouldn't be used directly.
Instead, use the specialized classes that inherits from it.
}
\section{Inherits}{

\describe{
  \item{\code{\link{factor_estimate}}}{}
}
}
\examples{
fe1 <- factor_estimate_composite$new()
}
\keyword{data}

