% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_estimate_composite_custom.R
\docType{class}
\name{factor_estimate_composite_custsom}
\alias{factor_estimate_composite_custsom}
\title{factor_estimate_composite_custom}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
factor_estimate_composite_custsom
}
\value{
An instance of the factor_estimate_composite \code{\link{R6Class}}.
}
\description{
This class represents a risk model "composite factor estimate with custom logic".\cr
A "composite factor estimate with custom logic" is a factor estimation method that is based on:\cr
 \itemize{
  \item a set of \code{dependent_factors} (aka sub-factors),
  \item a \code{random_function} that applies arbitrary logic on the input \code{dependent_factors} to model the factor's random variable.
  }
The \code{dependent_factors} may in their turn be estimated with arbitrary estimation methods.\cr
}
\section{Inherits}{

\describe{
  \item{\code{\link{factor_estimate}}}{}
}
}
\examples{
fe1 <- factor_estimate_composite_custom$new()
}
\keyword{data}

