% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_estimate_gld.R
\docType{data}
\name{factor_estimate_gld}
\alias{factor_estimate_gld}
\title{An abstract class for a GLD-based factor estimates.
Subclasses may inherit from it to implement various estimation techniques,
such as the 3 points estimate which is going to be the first
implementation.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
factor_estimate_gld
}
\description{
For the time being, I only support the FMKH parameterization. If
other parameterizations become necessary in the future, these will
require specifically dedicated R6 classes because the way we tweak
lambda parameters here is strongly linked to FMKH logic.
}
\details{
TODO:
- data validation and general help to friendly manage lambda parameters
}
\keyword{datasets}

