% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_estimate_gld_3points.R
\docType{class}
\name{factor_estimate_gld_3points}
\alias{factor_estimate_gld_3points}
\title{factor_estimate_gld_3points}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
factor_estimate_gld_3points
}
\value{
An instance of the \code{factor_estimate_gld_3points} \code{\link{R6Class}}.
}
\description{
A PERT-like factor estimate based on the Generalized Lambda (aka Tukey) Distribution.\cr
}
\details{
For the time being, I only support GLD's FMKH parameterization.
If other parameterizations become necessary / interesting in the future,
these will require specifically dedicated R6 classes,
because the way we tweak lambda parameters here is strongly linked to FMKH logic.\cr

Future enhancements:
- Make lambda parameters read-only for class users
  in such a way as to guarantee consistency between
  the fitted distribution and the estimation parameters.
}
\section{Fields}{

\describe{
\item{\code{estimated_range_min_value}}{The lower value of the 3 points estimate.}

\item{\code{estimated_mode_value}}{The "typical" value of the 3 points estimate.}

\item{\code{estimated_range_max_value}}{The upper value of the 3 points estimate.}

\item{\code{estimated_range_size_proba}}{Default: .9. Possible values: 0 < x < 1. The size of the probabilistic range estimate. The default .9 leaves .05 on both sides of the distribution.}

\item{\code{limit_min_value}}{A strict lower limit. 0 if you don't consider positive risks.}

\item{\code{limit_max_value}}{A strict upper limit. May be the company stock value if you don't consider externalities.}
}}
\section{Inherits}{

\describe{
  \item{\code{\link{factor_estimate_gld}}}{}
}
}

\section{Methods}{

\describe{
  \item{\code{new(estimated_range_min_value, estimated_mode_value, estimated_range_max_value, estimated_range_size_proba, limit_min_value, limit_max_value)}}{Create a new object of this class.}
  \item{\code{plot_density()}}{Plot the PDF.}
}
}
\examples{
fe1 <- factor_estimate_gld_3points$new(
  estimated_range_min_value = 10,
  estimated_mode_value = 90,
  estimated_range_max_value = 100,
  estimated_range_size_proba = .9,
  limit_min_value = 0,
  limit_max_value = 200)
fe1$plot_density()
}
\keyword{data}

