% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_estimate_poisson.R
\docType{data}
\name{factor_estimate_poisson}
\alias{factor_estimate_poisson}
\title{factor_estimate_poisson}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
factor_estimate_poisson
}
\description{
An abstract class for a Poisson-based factor estimates.
Subclasses may inherit from it to implement various estimation techniques,
such as the 3 points estimate which is the first planned implementation.
}
\keyword{datasets}

