% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_estimate_poisson_mode.R
\docType{data}
\name{factor_estimate_poisson_mode}
\alias{factor_estimate_poisson_mode}
\title{Poisson Range (2 points) estimate}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
factor_estimate_poisson_mode
}
\description{
Poisson Range (2 points) estimate
}
\keyword{datasets}

