% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_estimate_poisson_range.R
\docType{data}
\name{factor_estimate_poisson_range}
\alias{factor_estimate_poisson_range}
\title{Poisson Range (2 points) estimate}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
factor_estimate_poisson_range
}
\description{
Poisson Range (2 points) estimate
}
\keyword{datasets}

