% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_factor.r
\docType{data}
\name{fair_factor}
\alias{fair_factor}
\title{FAIR factors}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
fair_factor
}
\arguments{
\item{method2order}{method to order colors (\code{"hsv"} or \code{"cluster"})}

\item{cex}{character expansion for the text}

\item{mar}{margin paramaters; vector of length 4 (see \code{\link[graphics]{par}})}
}
\value{
None

A vector of probabilities

A vector of quantiles
}
\description{
Creates a plot of the crayon colors in \code{\link{brocolors}}
}
\examples{
plot_crayons()

}
\keyword{datasets}

