% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_function_1param_match.R
\name{find_function_1param_value_divideby2}
\alias{find_function_1param_value_divideby2}
\title{find_function_1param_value_divideby2}
\usage{
find_function_1param_value_divideby2(f, y_target_value, x_first_guess = NULL,
  x_first_step = NULL, x_search_limit_min = NULL,
  x_search_limit_max = NULL, y_precision = NULL, max_iteration = NULL,
  verbosity = NULL)
}
\arguments{
\item{f}{A function receiving a numeric value as its first parameter followed by ...}
}
\value{
nicely formatted numbers
}
\description{
Given \code{f(x) = y} where \code{f} is a monotonic function, find \code{x} for a given \code{y}.
\cr Use a naive iterative approach dividing an initial step by 2 until a solution is found within desired precision.
\cr Non-monotic functions are supported when the search range is limited to a monotonic subset of the function's range.
}
\examples{
fn(c(10.555,3.14),2)

}

