% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gld_3points.R
\name{fit_gld_3points}
\alias{fit_gld_3points}
\title{fit_gld_3points}
\usage{
fit_gld_3points(estimated_range_min_value = NULL,
  estimated_mode_value = NULL, estimated_range_max_value = NULL,
  estimated_range_size_proba = NULL, verbosity = NULL,
  max_iteration = NULL, precision = NULL, ...)
}
\arguments{
\item{estimated_range_min_value}{The lower value of the 3 points estimate.}

\item{estimated_mode_value}{The mode or "typical" value of the 3 points estimate.}

\item{estimated_range_max_value}{The upper value of the 3 points estimate.}

\item{estimated_range_size_proba}{Default: .9. Possible values: 0 < x < 1. The size of the probabilistic range estimate. The default .9 leaves .05 on both sides of the distribution.}
}
\value{
the lambda parameters of the generalized lambda distribution that yields the estimated 3 points.
}
\description{
Part of the fit_[distribution]_[method] function series.
Finds the lambda parameters of the generalized lambda distribution (aka Tukey distribution) that best matches the input 3 points estimate.
}
\examples{
xxxx(estimated_mode = 17.35)

}

