% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gld_3points_lambda1.R
\name{fit_gld_3points_lambda1}
\alias{fit_gld_3points_lambda1}
\title{fit_gld_3points_lambda1}
\usage{
fit_gld_3points_lambda1(lambda1, lambda2, lambda3, lambda4,
  estimated_range_min_value, estimated_mode_value, estimated_range_max_value,
  verbosity = NULL, ...)
}
\arguments{
\item{estimated_range_min_value}{The lower value of the 3 points estimate.}

\item{estimated_mode_value}{The mode or "typical" value of the 3 points estimate.}

\item{estimated_range_max_value}{The upper value of the 3 points estimate.}
}
\value{
A numeric vector of size 1 representing an optimized lambda1 (location) parameter that further fits a 3 points estimate.
}
\description{
This function optimizes the lambda1 (location) parameter of a GLD distribution to further fit a 3 points estimate.
\cr This is an "internal" function called by fit_gld_3points.
}

