% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gld_3points_lambda2.R
\name{fit_gld_3points_lambda2}
\alias{fit_gld_3points_lambda2}
\title{fit_gld_3points_lambda2}
\usage{
fit_gld_3points_lambda2(estimated_range_min_value, estimated_mode_value,
  estimated_range_max_value, estimated_range_size_proba, scaling_side = NULL,
  verbosity = NULL, ...)
}
\arguments{
\item{estimated_range_min_value}{The lower value of the 3 points estimate.}

\item{estimated_mode_value}{The mode or "typical" value of the 3 points estimate.}

\item{estimated_range_max_value}{The upper value of the 3 points estimate.}

\item{estimated_range_size_proba}{Default: .9. Possible values: 0 < x < 1. The size of the probabilistic range estimate. The default .9 leaves .05 on both sides of the distribution.}
}
\value{
A numeric vector of size 1 representing an optimized lambda2 (scale) parameter that further fits a 3 points estimate.
}
\description{
This function optimizes the lambda2 (scale) parameter of a GLD distribution to further fit a 3 points estimate.
\cr This is an "internal" function called by fit_gld_3points.
}

