% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gld_3points_lambda4.R
\name{fit_gld_3points_lambda4}
\alias{fit_gld_3points_lambda4}
\title{fit_gld_3points_lambda4}
\usage{
fit_gld_3points_lambda4(lambda1, lambda2, lambda3, lambda4,
  estimated_range_max_value, estimated_range_max_proba, verbosity = NULL, ...)
}
\arguments{
\item{estimated_range_max_value}{The upper value of the 3 points estimate.}
}
\value{
A numeric vector of size 1 representing an optimized lambda4 (shape) parameter that further fits a 3 points estimate.
}
\description{
This function optimizes the lambda4 (shape) parameter of a GLD distribution to further fit a 3 points estimate.
\cr This is an "internal" function called by fit_gld_3points.
}

