% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gld_3points.R
\name{fit_gld_3points_location}
\alias{fit_gld_3points_location}
\title{fit_gld_3points_location}
\usage{
fit_gld_3points_location(lambda1, lambda2, lambda3, lambda4,
  estimated_range_min_value, estimated_mode_value, estimated_range_max_value,
  verbosity = NULL, ...)
}
\description{
"Internal" function called by fit_gld_3points to fit the location parameter.
}

