% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gld_3points.R
\name{fit_gld_3points_scale}
\alias{fit_gld_3points_scale}
\title{fit_gld_3points_scale}
\usage{
fit_gld_3points_scale(estimated_range_min_value, estimated_mode_value,
  estimated_range_max_value, estimated_range_size_proba, scaling_side = NULL,
  verbosity = NULL, ...)
}
\description{
"Internal" function called by fit_gld_3points to fit the scale parameter.
}

