% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gld_3points.R
\name{fit_gld_3points_skew_left}
\alias{fit_gld_3points_skew_left}
\title{fit_gld_3points_skew_left}
\usage{
fit_gld_3points_skew_left(lambda1, lambda2, lambda4, estimated_range_min_value,
  estimated_range_min_proba, verbosity = NULL, ...)
}
\description{
"Internal" function called by fit_gld_3points to fit the lambda3 parameter.
}

