% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gld_3points.R
\name{fit_gld_3points_skew_right}
\alias{fit_gld_3points_skew_right}
\title{fit_gld_3points_skew_right}
\usage{
fit_gld_3points_skew_right(lambda1, lambda2, lambda3, estimated_range_max_value,
  estimated_range_max_proba, verbosity = NULL, ...)
}
\description{
"Internal" function called by fit_gld_3points to fit the lambda4 parameter.
}

