% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gpd_range.R
\name{fit_gpd_range}
\alias{fit_gpd_range}
\title{fit_gpd_range}
\usage{
fit_gpd_range(estimated_mode = NULL, estimated_range_max_value = NULL,
  estimated_range_size_proba = NULL, verbosity = NULL, ...)
}
\arguments{
\item{search_range_start}{the minimal value of the range where the mode will be searched}

\item{search_range_end}{the maximal value of the range where the mode will be searched}
}
\value{
the estimated parameters
}
\description{
Finds the lambda and theta parameters of a generalized poisson distribution
that best matches the desired range quantiles.
}
\examples{
N/A

}

