% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{fn}
\alias{fn}
\title{fn = format number}
\usage{
fn(x, digits = NULL)
}
\arguments{
\item{x}{a vector of numbers}

\item{digits}{the number of digits to be displayed}
}
\value{
nicely formatted numbers
}
\description{
A shortcut function to make the string manipulation strings less verbose.
TODO: This function is very procedural and is probably inefficient performance wise. Rewrite it.
}
\examples{
fn(c(10.555,3.14),2)

}

