% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R, R/format_number.R
\name{fn}
\alias{fn}
\alias{format_number}
\title{format_number, fn}
\usage{
fn(x, digits = NULL, ...)

format_number(x, digits = NULL, ...)
}
\arguments{
\item{x}{a vector of numbers}

\item{digits}{the number of digits to be displayed}
}
\value{
nicely formatted numbers
}
\description{
A friendly function to simplify the output of nicely formatted numbers.
\describe{
  \item{\code{format_number}}{The original function name.}
  \item{\code{fn}}{A shortcut function alias.}
}
\cr \strong{TODO}: This function is a bit procedural and is probably not optimal from a in terms of performances. We should test it with large volumes of numbers and optimize it as needed.
}
\examples{
format_number(c(10.555,3.14),2)
fn(c(10.555,3.14),2)

}

