% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist_mode_from_pdf.R
\name{get_dist_mode_from_pdf}
\alias{get_dist_mode_from_pdf}
\title{get_dist_mode_from_pdf}
\usage{
get_dist_mode_from_pdf(pdf, search_range_start, search_range_end,
  tolerance = NULL, verbosity = NULL, ...)
}
\arguments{
\item{pdf}{the input probability density function (PDF)}

\item{search_range_start}{the minimal value of the range where the mode will be searched}

\item{search_range_end}{the maximal value of the range where the mode will be searched}
}
\value{
the estimated mode
}
\description{
Shortcut function to find the mode of an arbitrary probability density function (PDF).
Uses optimization, i.e. may lead to imprecise results with exotic input functions.
It is helpful when an analytic solution is not handily available.
}
\details{
Works only for continuous distribution functions.
}
\examples{
get_dist_mode_from_pdf(function(x){return(dnorm(x = x, mean = 17, sd = 5))}, -100, 100)

}

