% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist_type.R
\name{get_dist_type}
\alias{get_dist_type}
\title{get_dist_type}
\usage{
get_dist_type(dist_name, ...)
}
\arguments{
\item{dist_name}{the shortened or complete name of the distribution.}
}
\value{
"discrete", "continuous", or an error if the distribution is unknown.
}
\description{
Informs us if a distribution is discrete or continuous.
}
\details{
Current development:
As of today, the list of distributions supported by the function
is far from satisfactory. I just implemented a few names that I
immediately needed + a few that I grabbed from wikipedia below.

Future improvements:
- Provide support for all well-known distributions.
- Find a more elegant way to manage distribution aliases, shortened versions, etc.

References:
- List of probability distributions, Wikipedia
  https://en.wikipedia.org/wiki/List_of_probability_distributions
}
\examples{
get_dist_type(dist_name="norm")
get_dist_type(dist_name="pois")

}

