% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gld_mode.R
\name{get_gld_mode}
\alias{get_gld_mode}
\title{get_gld_mode}
\usage{
get_gld_mode(lambda1 = NULL, lambda2 = NULL, lambda3 = NULL,
  lambda4 = NULL, search_range_start = NULL, search_range_end = NULL, ...)
}
\arguments{
\item{lambda1}{}

\item{lambda2}{}

\item{lambda3}{}

\item{lambda4}{}

\item{search_range_start}{}

\item{search_range_end}{}
}
\value{
The mode.
}
\description{
Returns the mode of the GLD (Generalized Lambda Distribution), aka Tukey distribution.
\cr When the shape parameters \code{lambda3 = -1, lambda4 = -1}, the mode is equivalent to the location parameter \code{lambda1}.
\cr But when the shape parameters are tweaked, we observe a shift of the distribution mode.
\cr This behavior is accentuated when \code{lambda2} approaches 0.
}
\section{Future improvements}{

\itemize{
  \item{ Perform some research to check if there exists an analytical solution to this function. Then, implement it to avoid the usage of optimization. In its current version, the function relies on optimization to find a good estimate for the mode. Consequently, the precision of the mode or computational intensity to find it are not optimal. }
  \item{ If no analytical solution is found and if \code{search_range_start} and \code{search_range_end} are not provided, provide best effort default values. }
}
}
\examples{
get_gld_mode(lambda1 = 100, lambda2 = 100, lambda3 = -10, lambda4 = -1, search_range_start = -1000, search_range_end = 1000)

}

