% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_nanull.R
\name{is_nanull}
\alias{is_nanull}
\title{is_nanull}
\usage{
is_nanull(x, ...)
}
\arguments{
\item{x}{a variable}
}
\value{
boolean
}
\description{
A shortcut function that test if a variable \code{x} is \code{NULL} or \code{NA}.
If \code{x} is a vector, \code{is_nanull} does not test the elements of the vector,
but returns \code{FALSE} because a vector is an object that is neither \code{NULL} nor \code{NA}.
}
\examples{
is_nanull(NULL)
is_nanull(NA)
is_nanull(0)
is_nanull(c(NA,NA,NA))

}

