% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_void.R
\name{is_void}
\alias{is_void}
\title{is_void}
\usage{
is_void(x, ...)
}
\arguments{
\item{x}{a variable}
}
\value{
boolean
}
\description{
Systematically testing for variable "emptiness" in R is a bit of a pain.
This is a shortcut function that tests if a variable \code{x} is \code{NULL} or \code{NA}.
If \code{x} is a vector, \code{is_void} does not test the elements of the vector,
but returns \code{FALSE} because a vector is an object that is neither \code{NULL} nor \code{NA}.
}
\examples{
is_void(NULL)
is_void(NA)
is_void(0)
is_void(c(NA,NA,NA))

}

