% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximize_function.R
\name{maximize_function}
\alias{maximize_function}
\title{maximize_function}
\usage{
maximize_function(f = NULL, search_range_start = NULL,
  search_range_end = NULL, tolerance = NULL, verbosity = NULL, ...)
}
\arguments{
\item{f}{a function accepting a single parameter.}

\item{search_range_start}{the minimal value of the range where the mode will be searched}

\item{search_range_end}{the maximal value of the range where the mode will be searched}

\item{verbosity}{0: no output (default), >0: more output}
}
\value{
the guessed parameter value that maximizes the function output
}
\description{
A friendly wrapper to find a function's max.
}
\examples{
maximize_function(f = function(x) { return(1 / (x - 17)) }, search_range_start = -100, search_range_end = 100, verbosity = 1)

}

