% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\docType{data}
\name{measure}
\alias{measure}
\title{measure}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
measure
}
\arguments{
\item{value}{The real value of the measure (i.e. the number).}

\item{unit}{The unit of measurement.}
}
\description{
Represent a real-value accompanied with a unit of measurement.
\cr In future versions, this class may be extended to provide support for unit conversions.
}
\examples{
m1 <- measure$new(value = 100, unit = "year")

}
\keyword{datasets}

