% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimax.R
\name{mimax}
\alias{mimax}
\title{mimax}
\usage{
mimax(x, minimum = NULL, maximum = NULL, ...)
}
\arguments{
\item{x}{A vector of numeric values.}

\item{minimum}{The minimal allowed value for \code{x}. If \code{NULL}, no minimum will be applied.}

\item{maximum}{The maximal allowed value for \code{x}. If \code{NULL}, no maximum will be applied.}
}
\value{
Numeric vector. Bounded x.
}
\description{
A shortcut function that limits a value between an inclusive minimum and an inclusive maximum.
}
\examples{
mimax(seq(-10,10,.1),-2,2)

}

