% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimize_function.R
\name{minimize_function}
\alias{minimize_function}
\title{minimize_function}
\usage{
minimize_function(f = NULL, search_range_start = NULL,
  search_range_end = NULL, tolerance = NULL, verbosity = NULL, ...)
}
\arguments{
\item{f}{a function accepting a single parameter.}

\item{search_range_start}{the minimal value of the range where the mode will be searched}

\item{search_range_end}{the maximal value of the range where the mode will be searched}

\item{verbosity}{0: no output (default), >0: more output}
}
\value{
the guessed parameter value that minimizes the function output
}
\description{
A friendly wrapper to find a function's min
}
\examples{
minimize_function(f = function(x) { return(1 / (x - 17)) }, search_range_start = -100, search_range_end = 100, verbosity = 1)

}

