% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_estimate.R
\docType{data}
\name{model_estimate}
\alias{model_estimate}
\title{model_estimate}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
model_estimate
}
\arguments{
\item{name}{the unique name used to identify this factor within its model}

\item{dist}{the theoretical probability distribution of this factor}
}
\value{
an instance of a mode factor
}
\description{
An abstract class that is used as a glueing interface between
the model_analysis + model_factor data model on one hand,
and an extensible set of estimation methods or approaches.
}
\examples{
f1 <- model_factor$new(name = "x", dist = "lnorm")

}
\keyword{datasets}

