% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_estimate_gldpert.R
\docType{data}
\name{model_estimate_gldpert}
\alias{model_estimate_gldpert}
\title{model_estimate_gldpert}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
model_estimate_gldpert
}
\arguments{
\item{xxx}{blablabla}
}
\value{
an instance of a GLD-PERT estimation
}
\description{
The GLD-PERT estimation method is based on the PERT estimation style,
well adapted for subject matter expert estimates.
It is primarily based on 3 estimates points (quantiles)
and a range size.
The range size is defaulted to .9, leaving .05 on each side
of the distribution.
These parameters are then fitted to a GLD (Generalized Lambda Distribution)
distribution. GLD was chosen for its great flexibility, fitting estimated
points in most circumstances. In addition to this, many usual distributions
may be represented by a GLD given the proper lambda parameters.
}
\examples{
f1 <- model_factor$new(name = "x", dist = "lnorm")

}
\keyword{datasets}

