% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_factor.R
\docType{class}
\name{model_factor}
\alias{model_factor}
\title{model_factor}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
model_factor
}
\value{
An instance of the \code{model_factor} \code{\link{R6Class}}.
}
\description{
A risk model factor.
}
\section{Inherits}{

\describe{
  \item{This is a root class.}{}
}
}

\section{Methods}{

\describe{
  \item{get_random(n = 1, output_class = "vector")}{ Returns a random sample of size \code{n}. Returns a vector by default. If \code{output_class} = "data.frame", returns a data.frame with a column "factor_value". This second parameterization may be enriched by R6 subclasses to provide additional columns with complementary information. }
}
}
\examples{
f1 <- model_factor$new()
}
\keyword{data}

