% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overplot_estimates.R
\name{overplot_estimates}
\alias{overplot_estimates}
\title{overplot_estimates}
\usage{
overplot_estimates(graph, estim_quantiles, estim_labels, ...)
}
\arguments{
\item{graph}{a ggplot graph, typically a PDF or CDF}

\item{estim_quantiles}{a vector of estimated quantiles (aka x values)}

\item{estim_labels}{meaningful labels to describe the estimated quantiles}
}
\value{
an good looking graph enriched with visual representations for the estimates
}
\description{
Enriches a graph with good looking vertical interects representing estimates.
}
\examples{
fun <- function(x, ...) { return(dlnorm(x = x, meanlog = 100, sdlog = 20, log = FALSE))  }
graph <- plot_probability_density_function(fun = fun, x_start = 10, x_end = 150)
graph_enriched <- overplot_estimates(graph=graph, estim_quantiles = c(50,90), estim_labels = c("a","b"))
graph_enriched

}

