% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overplot_horizontal_lines.R
\name{overplot_horizontal_lines}
\alias{overplot_horizontal_lines}
\title{overplot_vertical_lines}
\usage{
overplot_horizontal_lines(graph, y_values, y_labels = NULL, color = "black",
  alpha = 1, ...)
}
\arguments{
\item{graph}{a ggplot graph, typically a PDF or CDF}

\item{estim_quantiles}{a vector of estimated quantiles (aka x values)}

\item{estim_labels}{meaningful labels to describe the estimated quantiles}
}
\value{
an good looking graph enriched with visual representations for the estimates
}
\description{
Enriches a graph with good looking vertical lines.
For instance, this may be used to represend point estimates and limits in a PDF.
}
\examples{
fun <- function(x, ...) { return(dlnorm(x = x, meanlog = 100, sdlog = 20, log = FALSE))  }
graph <- plot_probability_density_function(fun = fun, x_start = 10, x_end = 150)
graph_enriched <- overplot_horizontal_lines(graph=graph, x_values = c(.05,.95), x_labels = c("a","b"))
graph_enriched

}

