% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cumulative_distribution_function.R
\name{plot_cumulative_distribution_function}
\alias{plot_cumulative_distribution_function}
\title{plot_cumulative_distribution_function}
\usage{
plot_cumulative_distribution_function(fun, x_start, x_end, ...)
}
\arguments{
\item{fun}{a parameterized cumulative distribution function receiving a single x parameter}

\item{x_start}{the left most position that will be displayed on the x axis}

\item{x_end}{the right most position that will be displayed on the x axis}
}
\value{
a good looking graph
}
\description{
Produces a good looking graph of a cumulative distribution function.
}
\examples{
fun <- function(x, ...) { return(dlnorm(x = x, meanlog = 100, sdlog = 20, log = FALSE))  }
graph <- plot_cumulative_distribution_function(fun = fun, x_start = 10, x_end = 150)
graph

}

