% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factor.R
\name{plot_factor}
\alias{plot_factor}
\title{plot_factor}
\usage{
plot_factor(factor, ...)
}
\arguments{
\item{factor}{an instance of the model_factor R6 class}
}
\value{
an good looking graph enriched with visual representations for the estimates
}
\description{
Produces a good looking multi-graph visual representation of a model factor.
}
\examples{
f1 <- model_factor$new(name = "impact sample")
f1$estim_3_points_betapert(
 estim_min = 200,
 estim_typical = 1300,
 estim_max = 1500,
 range_size = .9)
plot_factor(f1)

}

