% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_probability_mass_function.R
\name{plot_probability_mass_function}
\alias{plot_probability_mass_function}
\title{plot_probability_mass_function}
\usage{
plot_probability_mass_function(fun, x_start, x_end, ...)
}
\arguments{
\item{fun}{a parameterized probability mass function receiving a single x parameter}

\item{x_start}{the left most position that will be displayed on the x axis}

\item{x_end}{the right most position that will be displayed on the x axis}
}
\value{
a good looking graph
}
\description{
Generates a good looking graph representing a probability mass function.
}
\examples{
fun <- function(x, ...) { return(dpois(x = x, lambda = 5, log = FALSE))  }
graph <- plot_probability_mass_function(fun = fun, x_start = 0, x_end = 20)
graph

}

