% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample.R
\name{plot_sample}
\alias{plot_sample}
\title{plot_sample}
\usage{
plot_sample(sample, title = NULL, subtitle = NULL, caption = NULL,
  x_start = NULL, x_end = NULL, bins = NULL, x_scale_type = NULL,
  y_scale_type = NULL, variable_type = NULL, ...)
}
\arguments{
\item{sample}{A vector of the population sample.}

\item{title}{A title for the plot.}

\item{subtitle}{A subtitle for the plot. Unfortunately, this is not yet supported with the non-dev version of GGPlot2, but should come soon.}

\item{caption}{A caption for the plot. Unfortunately, this is not yet supported with the non-dev version of GGPlot2, but should come soon.}

\item{x_start}{The left most position that will be displayed on the x axis. Overflowing values will be ignored.}

\item{x_end}{The right most position that will be displayed on the x axis. Overflowing values will be ignored.}

\item{bins}{The number of bins in the histogram.}

\item{x_scale_type}{"Normal" (default) or "Log10".}

\item{y_scale_type}{"Normal" (default) or "Log10".}

\item{variable_type}{"Discrete" (default) or "Continuous".}

\item{verbosity}{0: no messages. > 0 more and more verbose messages.}
}
\value{
a good looking graph
}
\description{
A friendly shortcut function that produces a good looking graph of a population sample.
The main and lower part is composed of an histogram.
The upper part is enriched with a marginal box plot.
}
\examples{
plot_sample(sample = rnorm(n = 10000), title = "Normal sample")
plot_sample(sample = rpois(n = 10000, lambda = exp(1)), title = "Poisson sample", variable_type = "Discrete")
plot_sample(sample = rgeom(n = 1000, p = .2), title = "Geometric sample", variable_type = "Discrete")
}

