% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vignette.R
\name{plot_vignette}
\alias{plot_vignette}
\title{plot_vignette}
\usage{
plot_vignette(title = NULL, subtitle = NULL, caption = NULL,
  text = NULL...)
}
\arguments{
\item{title}{the title of the plot}

\item{text_lines}{a vector of text lines to be included in the plot.}
}
\value{
a good looking graph
}
\description{
Produces a good looking plot of a summary text.
May be nicely integrated in a plot grid to provide summary stats.
}
\examples{
N/A

}

