% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_mass_function.R
\name{probability_mass_function}
\alias{probability_mass_function}
\title{probability_mass_function}
\usage{
probability_mass_function(dist_name, x, ...)
}
\arguments{
\item{x}{a vector of quantiles}
}
\value{
a vector of probabilities
}
\description{
This is a PMF wrapper function.
It receives the distribution name as a parameter, making the function distribution-dynamic.
Distribution parameters passed in ... are simply transmitted to the underlying distribution function.
}
\examples{
probability_mass_function('poissonpert',x=c(1,5,7), lambda=3)

}

