% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.R
\docType{data}
\name{rand}
\alias{rand}
\title{rand}
\format{An object of class \code{randomity} (inherits from \code{R6}) of length 5.}
\usage{
rand
}
\description{
A helper object to easily get pseudo random numbers with high entropy.
From time to time, you will observe that calls to this object's methods
take a bit more time than usual. This is due to the reseeding mechanism
that relies on external entropy.
This "global" object is an instance of the randomity class.
Type: help(randomity) for more information.
}
\examples{
print("Get a single pseudo random number between 0 and 1")
rand$get()
print("Get a dozen pseudo random numbers between -100 and 100")
rand$get(n = 12,min = -100,max = 100)
}
\keyword{datasets}

