% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomity.R
\docType{class}
\name{randomity}
\alias{randomity}
\title{randomity}
\format{\code{\link{R6Class}} object.}
\usage{
randomity
}
\value{
Object of \code{\link{R6Class}} with properties and methods for pseudo random number generation.
}
\description{
A helper class that generates pseudo random numbers with high entropy. Do not use this class directly: instead, use the global object \code{rand} that is already instanciated.
}
\section{Methods}{

\describe{
  \item{\code{new()}}{Create a new object of this class.}
  \item{\code{get(n,min,max,verbosity)}}{Generate \code{n} pseudo random numbers with values between \code{min} and \code{max}.}
}
}

\section{Future enhancements}{

\describe{
  \item{- Add entropy from local system time}{}
  \item{- Add entropy from local system information}{}
}
}
\examples{
test <- randomity$new()
test$get(n = 32, min = 0, max = 10)
}
\keyword{data}

