% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_plot.R
\name{sum_plot}
\alias{sum_plot}
\title{sum_plot}
\usage{
sum_plot(verbosity = NULL, ...)
}
\arguments{
\item{verbosity}{}

\item{...}{A series of n plots to be added together.}
}
\value{
An enriched plot.
}
\description{
Add together n plots.
\cr GGPlot plots may be enriched by adding together multiple geoms.
\cr But \code{NULL} and/or \code{NA} can't be added with geoms.
\cr This function makes it easier to add together a series of geoms, getting rid of \code{NULL} and \code{NA} in the process.
}
\examples{
sum_plot(p1, p2, p3)

}

