% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_pretty_print.R
\name{vector_pretty_print}
\alias{vector_pretty_print}
\title{vector_pretty_print}
\usage{
vector_pretty_print(v, ...)
}
\arguments{
\item{l}{the list}
}
\value{
a string that is a pretty representation of the vector
}
\description{
Converts a named vector to a single-line string pretty representation
}
\examples{
l <- list(1, 2, 3, 4)
names(l) <- c("a", "b", "c", "d")
print(list_pretty_print(l))

}

