% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_estimate.R
\docType{class}
\name{factor_estimate}
\alias{factor_estimate}
\title{factor_estimate}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
factor_estimate
}
\value{
An instance of the \code{factor_estimate} \code{\link{R6Class}}.
}
\description{
A risk model "factor estimate".
A "factor estimate" is an abstraction that separates the factor estimation method from the model factors.
This approach makes it possible to work separately on the risk model and on the best possible estimation methods to estimate the model factors.
\code{factor_estimate} is an "abstract" class that shouldn't be used directly.
Instead, use the specialized classes that inherits from it.
}
\section{Fields}{

\describe{
\item{\code{estimation_method_name}}{A descriptive name identifying the estimation method, e.g.: "PERT-like 3 points estimate".}

\item{\code{distribution_name}}{The name of the probability distribution, e.g. "Poisson", "Composite", etc.}

\item{\code{distribution_type}}{Either "Continuous" or "Discrete".}

\item{\code{limit_min_value}}{A strict lower bound applied to the factor simulation values. If NULL or NA, no lower bound will be applied.}

\item{\code{limit_max_value}}{A strict upper bound applied to the factor simulation values. If NULL or NA, no upper bound will be applied.}
}}
\section{Inherits}{

\describe{
  \item{This is a root class.}{}
}
}

\section{Methods}{

\describe{
  \item{get_random(n = 1, output_class = "vector")}{ Returns a random sample of size \code{n}. Returns a vector by default. If \code{output_class} = "data.frame", returns a data.frame with a column "factor_value". This second parameterization may be enriched by R6 subclasses to provide additional columns with complementary information. }
}
}
\examples{
fe1 <- factor_estimate$new()
}
\keyword{data}

