% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_estimate_gld_3points.R
\docType{data}
\name{factor_estimate_gld_3points}
\alias{factor_estimate_gld_3points}
\title{factor_estimate_gld_3points: GLD 3 points estimate}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
factor_estimate_gld_3points
}
\description{
For the time being, I only support GLD's FMKH parameterization.
If other parameterizations become necessary / interesting in the future,
these will require specifically dedicated R6 classes,
because the way we tweak lambda parameters here is strongly linked to FMKH logic.
}
\details{
Future enhancements:
- Make lambda parameters read-only for class users
  in such a way as to guarantee consistency between
  the fitted distribution and the estimation parameters.
}
\keyword{datasets}

