% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_estimate_poisson_mode.R
\docType{class}
\name{factor_estimate_poisson_mode}
\alias{factor_estimate_poisson_mode}
\title{factor_estimate_poisson_mode}
\format{\code{\link{R6Class}} object.}
\usage{
factor_estimate_poisson_mode
}
\value{
Object of \code{\link{R6Class}} with properties and methods for simple integration in a risk model.
}
\description{
A risk model factor estimate.
Based on a Poisson distribution, ideal for event frequency modelling.
The estimation is based on the distribution mode.
}
\section{Fields}{

\describe{
\item{\code{dist_mode}}{The mode of the fitted distribution.}
}}
\section{Methods}{

\describe{
  \item{\code{new(estimated_mode_value,time_interval_friendly_name)}}{This method is used to create a new object of this class with \code{estimated_mode_value} as the estimated mode.}
  \item{\code{plot_density()}}{Plot the PDF.}
}
}
\examples{
factor_estimate_poisson_mode$new(estimated_mode_value = 100)
}
\keyword{data}

