% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_poisson_mode.R
\name{fit_poisson_mode}
\alias{fit_poisson_mode}
\title{fit_poisson_mode}
\usage{
fit_poisson_mode(estimated_mode_value = NULL, verbosity = NULL, ...)
}
\arguments{
\item{estimated_mode_value}{the estimated mode of the unkwown distribution}
}
\value{
the lambda parameter of the poisson distribution that yields the estimated mode
}
\description{
Finds the lambda parameter of a poisson distribution that best matches the estimated mode.
This function is trivial but was written for consistency purposes with the other fitting functions.
}
\examples{
fit_poisson_lambda(estimated_mode = 17.35)

}

