% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqimpact.R
\name{freqimpact}
\alias{freqimpact}
\title{freqimpact}
\usage{
freqimpact(n, frequency_function, impact_function, output_class, ...)
}
\arguments{
\item{n}{The desired size of the frequency vector.}

\item{frequency_function}{A frequency function supporting parameter \code{n} that returns a vector of integer frequencies of size \code{n}.}

\item{impact_function}{An impact function supporting parameter \code{n} that returns a vector of impacts of size \code{n}.}

\item{output_class}{Default: "vector". If \code{output_class} = "data.frame", returns a detailed data.frame with columns \code{frequency}, \code{impact_list} and \code{factor_value}.}
}
\value{
a vector or data.frame
}
\description{
Blablabla
}
\examples{
freqimpact(
  n = 8,
  frequency_function = function(n) {return(rbinom(n = n, size = 4, prob = .5))},
  impact_function = function(n) { return(rnorm(n = n, mean = 0, sd = 100))},
  output_class = "data.frame")

}

