% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\docType{data}
\name{measure}
\alias{measure}
\title{measure}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
measure
}
\arguments{
\item{value}{the real value of the measure}

\item{unit}{the unit of measurement}
}
\description{
Represent a real-value accompanied with a unit of measurement.
}
\examples{
m1 <- measure$new(name = 100, unit = "year")

}
\keyword{datasets}

