% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_factor.R
\docType{class}
\name{model_factor}
\alias{model_factor}
\title{model_factor}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
model_factor
}
\value{
An instance of the \code{model_factor} \code{\link{R6Class}}.
}
\description{
A risk model factor.
}
\section{Inherits}{

\describe{
  \item{This is a root class.}{}
}
}
\examples{
f1 <- model_factor$new()
}
\keyword{data}

