% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_factor.R
\docType{data}
\name{model_factor}
\alias{model_factor}
\title{model_factor}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
model_factor
}
\arguments{
\item{name}{the unique name used to identify this factor within its model}

\item{dist}{the theoretical probability distribution of this factor}
}
\value{
an instance of a mode factor
}
\description{
Represent a component of a \code{\link{model}}
}
\examples{
f1 <- model_factor$new(name = "x", dist = "lnorm")

}
\keyword{datasets}

