% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.R
\docType{data}
\name{randomity}
\alias{randomity}
\title{randomity}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
randomity
}
\description{
A helper class that generates pseudo random numbers with high entropy.
}
\details{
Future enhancements:
- Add entropy from local system time
- Add entropy from local system information
}
\examples{
test <- randomity$new()
test$get(n = 32, min = 0, max = 10)

}
\keyword{datasets}

